Added 'domain' to list of terms
I noticed I was not successful in the final step of the pull request.  I wasn't seeing the change outside of my local repository.  I tried to back up and start over again, but I think I'm just making more of a mess.  I am changing the double-quotation marks I used in my first commit to single-quotation marks as I'm seeing elsewhere in case that is an issue.